Add Custom Initializer for HomeViewController
HomeViewController now has a custom initializer for initialzing its view model. This is preferable to the previous approach because it did not properly convey that HomeViewController depends on HomeViewModel. This change required altering the MainViewController protocol to reflect that the viewModel property is only gettable so it could exist as a constant.