Add LoginViewController
LoginViewController is a new UIViewController that will allow the user to sign into their account. It doesn't do anything yet (tapping Log In will bring the user to HomeViewController no matter what right now). The view fully supports Dynamic Type, with the exception of the passwordTextFieldButtons. It appears that a UITextField's rightView/leftView does not resize with dynamic type, so a workaround for this will be persued at a later time.