Finalize LoginView Design
LoginView's UITextFields now show the standard iOS clear button when text is present in either of them. This was done to simplify handling deleting their respective values from LoginViewModel. To facilitate this, the eye button for the password text field was moved outside of the text field. This prevents the need for worrying about how the clear button will interact with a the text field's rightView. Also, LoginView now supports Dynamic Type more thoroughly, with the exception of the password field's eye button. This button does not seem to resize, even with its adjustsImageSizeForAccessibilityContentSizeCategory set to true. This will be deferred for now.