Adopt the Coordinator Pattern
The Coordinator pattern has been adopted to decouple navigation code from all UIViewControllers. To facilitate this, OnboardingCoordinator manages navigation during onboarding (including signing in), HomeCoordinator manages navigation in the Home tab (although the UITabBar hasn't been added yet), and MainCoordinator manages switching between OnboardingCoordinator and HomeCoordinator. In this commit, OnboardingCoordinator is always first in the MainCoordinator's childCoordinators array. In a later commit, logic will be added to check whether or not a user is logged in before determining which Coordinator to use first.