Remove navigationController Parameter from Coordinator Protocol
The navigationController parameter has been removed from the Coordinator protocol because not all Coordinators need a navigation controller. In the case of TabBarCoordinator, it was receiving a navigation controller it wasn't using. That's because a UITabBarController should be used as an app's root view controller, so there's never a need for it to be embedded in a navigation controller. Also, all of the view controllers and coordinators  within the tab bar controller create their own navigation controller, as opposed to receiving one injected by a parent.