Add ViewState Management to LoginView
Created a new ViewState protocol to which various ViewState enums will conform. These enums well help update the UI of various views as the user performs different actions. Also, all view models were updated with @MainActor so that all updates occur on the Main Actor. Finally, the new FBAuthError type makes it easier to present errors to the user when they're trying to log in.