Set MainCoordinator's Navigation Controller to nil On Successful Log In
MainCoordinator's childOnboardingCoordinatorDidFinish method now sets the MainCoordinator's navigationController property to nil. This was done because, previously, logging the user in was only clearing out the array of view controllers in the navigation controller. Although this worked in practice, it was still unnecessarily keeping the navigation controller alive in memory. To facilitate this functionality, childTabBarCoordinatorDidFinish creates a new navigation controller for MainCoordinator. This is necessary so that, when the user logs out, they're taken back to a new navigation controller that displays LoginViewController.