Add @MainActor to All Coordinators, Create Unit Test Target, Start DatabaseService Mocking
A MockDatabaseService object was created to make testing database calls in tests easier. To facilitate this, a new databaseService property was added to all coordinators and view models, as DatabaseService can no longer be a singleton. DatabaseServiceProtocol was also created to make it easier to use dependency injection when testing view models. Lastly, all Coordinator objects are now tagged @MainActor. This is because view models (which are tagged @MainActor) now have to be initialized within coordinators since injecting a DatabaseServiceProtocol instance is required. An @MainActor object cannot be initialized in an object that is not also guaranteed to run on the Main Actor, so Coordinators must also be @MainActor.