Change LoginView Layout
LoginView now presents Sign Up options at the bottom of the view to better distribute UI elements across the screen. To facilitate Dynamic Type support for this new layout, loadSubviews was overridden to update constraints based on whether or not the new UIScrollView in LoginView needs to scroll. If it does, sign up elements will be displayed immediately under log in elements. If it does not need to scroll, then sign up options will be shown at the bottom of the UI.