Fix MainTextField and PasswordTextFieldStack Bugs
Fixed a bug where content was going underneath MainTextField's clear button. This was happening because the rects for the text within the text field were being adjusted without accounting for the rect of the clear button. Also, fixed a bug in PasswordTextFieldStack where typing a very long password would cause the text field to expand and push the eye button off the screen. This was fixed by setting the content compression resistance priority of the passwordTextField to .defaultLow for .horizontal.