Add Sign Up Functionality and Refactor Text Field Logic
A user can now sign up for an account in SignUpView, but logic is still missing (form validation, showing error messages, etc.). Also, PasswordTextFieldStackDelegate was deleted. This was done because having a separate delegate just for text fields within PasswordTextFieldStack was confusing, so PasswordTextFieldStack was refactored to allow for its parent view to perform UITextFieldDelegate actions.