Add MainCoordinator Bug Fix
MainCoordinator no longer has a UINavigationController in its initializer. This is because, when the user opens the app after they're already signed in, there is no need for MainCoordinator to own a navigation controller. As a result, MainCoordinator's navigationController property is now only assigned when the user opens the app when they aren't already signed in, and when the user logs out.