Refactor Error ViewState and Add More SignupViewModelTests
The error case of any ViewState enum no longer uses the Error type as its associated value. It uses a String that contains the error message instead. This was done because Error does not conform to equatable, so testing viewState property values of view models would've been challenging. With the String associated value, all ViewState enums automatically conform to Equatable and testing them is easier.