Convert NewTextEntry Files into AddEditTextEntry Files
NewTextEntry files are now AddEditTextEntry files so that the view can be used for both creating and updating text entries. AddEditNewTextEntryViewController is displayed when the user taps an entry in EntriesView. The save button in AddEditTextEntryViewController does not work as expected yet, though. To make this functionality possible, AddEditTextEntryViewCoordinator was created. This was necessary because AddEditTextEntryViewController needs to be presented via HomeViewController and EntriesViewController. So, if AddEditTextEntryViewController doesn't have its own coordinator type, then its coordinator property will sometimes refer to an unexpected coordinator (for example, HomeCoordinator could be referenced when AddEditTextEntryViewController is actually being presented from EntriesCoordinator).