Add Current User Dependency Injection
The logged in User object is now passed throughout the app via dependency injection. This was done to both make testing easier, make the codebase more expressive, and to avoid having to handle throughout the app to access the current user's UID, email address, etc. Since this requires an asynchronous task at startup to fetch the current user's data from Firestore, a StartupViewController was created. This view is now the rootViewController at startup to prevent the screen from going back in between presenting the LaunchScreen and the first view within the app.