Fix AddEditTextEntryViewController Bug and Refresh EntriesView
AddEditTextEntryViewController's coordinator was not configured to present an error when necessary. Now it is, so the view controller will present an error message when expected. Additionally, AddEditTextEntryViewController will now hide the save button when a TextEntry is being edited and all of its text is erased. Finally, EntriesViewController now uses viewIsAppearing to fetch all TextEntry objects. This was done so that the user sees the most up to date info every time the view is loaded.