Remove Elements from HomeView to Prepare for Re-Design
Most elements have been removed from HomeView in preparation for the re-design that will be integrated in a future commit. Elements that are no longer in use are now in the Old and Unused folder, as they may be re-implemented in similar ways throughout the app in a future commit.