Begin HomeView Redesign and Improve Dynamic Type Support
HomeView now shows HomeActivityStreakSection, which is hard coded with values at the moment but will include the user's current Entry streak. Additionally, Dynamic Type support has been improved in that the largest font sizes are now displayed properly. Before, the biggest font sizes were far too big. This was occurring because of lines such as this in UIFont+Helpers: . The problem with this line is that the value in the __TEXT	__DATA	__OBJC	others	dec	hex property is not hard coded; it will change every time the user's font size changes. As a result, when this font is used as a  in a  object,  uses an increasingly large value to determine font size for larger and larger fonts.