Update SignUpView with New Design, Fix Retain Cycle in EntriesViewController
SignUpView now shows the new design that has been implemented in other sections of the app. Also, the UIAction handlers in EntiresViewController now weakly reference self so that a retain cycle does not occur with the view. A retain cycle shouldn't occur anyway since the view controller will usually be around for the lifetime of the app, but better safe than sorry.