Add Keyboard Dismissal to SignUpView and LoginView
The user can now tap in blank space to dismiss the keyboard in LoginView and SignUpView. This renders the scroll view in LoginView unnecessary, as it can not scroll at any iPhone screen size or any Dynamic Type font. However, the scroll view hasn't been removed because it may be needed in the future.