Add Functionality for 2 Buttons to CustomAlert
CustomAlert can now display two buttons at one time. As of now, this functionality is only utilized when presenting a CustomAlert with a CustomAlertType of .confirmation. There is a known issue where the AutoLayout engine produces conflicting constraint error messages when the user changes the font size while an alert is being displayed. This will be deferred for now, as I can't figure out the issue, it's an unlikely edge case, and the broken constraints do not produce any strange visual issues.