Refactor Code that Runs When Font Changes in HomeView
In each of HomeView's sections, the method that's called when the font size changes is now called adjustLayoutForNewPreferredContentSizeCategory. The code in each of these methods was refactored into computed properties so that adjustLayoutForNewPreferredContentSizeCategory wouldn't need to be called in each section's init. This makes the code more expressive and readable, as the purpose of the adjustLayoutForNewPreferredContentSizeCategory is clearer.