Fix Cancelling Video Picking Bug
The user no longer sees an error when they cancel video picking in CreateVideoEntryViewController. This was occurring because the picker's delegate was throwing an error if the user finished picking without picking exactly 1 video. So, if they didn't pick any video, it was throwing an error. Now an error is not thrown unless the user somehow picks more than one video, which should never be possible.