Change Style and Layout for Buttons in CreateVideoEntryViewController
The buttons in CreateVideoEntryViewController now use the .primaryElement and .background colors so that they don't appear transparent and hard to see. The switch camera button was also moved to the bottom trailing edge to make the layout more similar to that of the stock Apple Camera app. Finally, the image for the button to show the video picker was changed to more closely match the style of the rest of the button images in the app, and all buttons now disappear when video recording starts.