Add UI for Uploading and Saving Video Entry to Device
There is now an upload button in UploadVideoView for uploading video entries, but this button doesn't do anything yet. Similarly, there's a toggle in UploadVideoView that the user can use to decide whether or not they want to save their video entry to their device before uploading, but this also doesn't do anything yet.