Rename Coordinator's Method for Showing Errors
The Coordinator protocol for presenting error messages was requiring a reference to the view controller that the error message was to be presented on. This is no longer necessary due to the way that AlertPresenter was refactored in a previous commit, so Coordinator now has a method called viewControllerShouldPresentErrorMessage(_:) that does the same thing as before without the unnecessary reference.