Confirm When Dismissing UploadVideoViewController, Switch to UIImagePickerController
CreateVideoEntryViewController now uses a UIImagePickerController instead of a PHPickerViewController because of its functionality for restricting selected video length. PHPickerViewController does not prompt the user to trim the selected video to the necessary length, nor can it filter videos based on length. Videos can only be filtered by length after they're loaded from the URL at which they're stored, which takes far too long for long videos. UIImagePickerController prompts the user to trim the video if they pick one longer than 5 minutes, so this is a much faster workflow. Additionally, the user is now asked to confirm if they wish to discard their video entry before they press the new BackButtonView in UploadVideoViewController.