Implement Saving Recorded Video to Photos Library
The UI in UploadVideoEntryViewController now works and saves the user's recorded video to their Photos library when they enable the toggle to do so. A UIProgressView appears during saving, and the user is not shown the option to save when they select a video entry from their Photos library. This is not necessary because they already have the video in their library, so there's no need to save it again.