Refactor and Add Bug Fixes to UploadVideoEntryView
dismissMediaControls() in UploadVideoEntryView now invalidates the hideMediaControlsTimer and sets it to nil. This is done because it prevents media controls from disappearing too quickly in the event that the user presents media controls, dismisses them before the timer ends, and then makes them reappear before the timer ends. Now, media controls will stay visible for the same period of time no matter when they're shown. Also code was refactored in UploadVideoEntryView to make it more readable.