Refactor CustomAlertView and Add Inadequate Permissions Message in CreateVideoEntryViewController
CustomAlertView has been made more robust to make it easier to perform different types of actions with its primary button. This was done to make it easier to add a CustomAlert to CreateVideoEntryViewController that tells the user when they don't have camera/microphone permissions, and then prompts them to open Settings to grant the permission(s). Additionally, changes were made to CustomAlert to make it more functional at higher dynamic type sizes. Specifically, top and bottom constraints were added so that long error messages don't push it off the screen.