Fix Bugs in UploadVideoEntryView
Scrubbing is now smoother in UploadVideoEntryView. This was done by adding a half a second tolerance in the view model's seekVideoPlayer(to:) method, which prevents the video player from having to do more work to ensure that the video is seeked exactly to where the user requested. Additionally, on touch down of the timeline slider, the video is now paused and the hideMediaControlsTimer is invalidated. This prevents the user from having the slider disappear or move while they're touching it.