Add Audio Recording and Playback Functionality
CreateVoiceEntryViewController now successfully records and plays back audio. Stop and play buttons are in place for testing purposes, better view state updates will be made in future commits. In this commit, pressing play before stopping causes a crash. This will be fixed in a future commit once the functionality is fully fleshed out.