Fix CreateVoiceEntryViewController Crash
CreateVoiceEntryViewModel was causing a crash in its deleteLocalRecording() method because it was being called regardless of whether or not an AVAudioRecorder had been created. The AVAudioRecorder in CreateVoiceEntryViewModel is being implicitly unwrapped, so this was causing a crash. Now the method checks to make sure the audio recorder is not nil before attempting to delete the locally recorded entry.