Update Error Presenting Functionality to Support Sheets
Prior to this commit, error alert presenting would put the alert behind a sheet if a sheet attempted to present an alert. This was happening because the alert was only being presented on top of the current UIWindow's root view controller's view, which gets covered by sheets. Now, the Coordinator error presenting method is named presentErrorMessage(onViewController:errorMessage:), as it accepts an optional parameter for the view controller on top of which the alert should be displayed. This parameter should be nil except for when a sheet is attempting to display an error. In that case, the sheet's navigation controller will be passed in so that its view can present the error directly. When this parameter is not nil, the alert is still presented on top of the current UIWindow's root view controller's view.