Add Toggling Between Incomplete and Complete Goals in GoalsViewController
The user can now toggle between their complete and incomplete goals in GoalsViewController to easily view sorted goals. Additionally, completing a goal will now trigger an animation on the tapped completion button and then move the completed goal to the completed goal table view. There is a known bug where completing the last incompletegoal and then creating a new goal immediately after causes the newly created goal to have a checkmark image instead of a circle image after it's created. It appears that somehow the table view is retaining the checkmark image from the previously completed goal and leaving it in place for the newly created goal. This bug does not occur if there are multiple incomplete goals remaining after completing one. For now, the workaround for this issue is to set each cell's complete goal button back to the circle image after the animation takes place. This prevents the table view from retaining the incorrect image for new goals, but it also can prevent the completion animation from taking place. This will be revisited later.