Add Functionality to WatchVideoEntryViewController
A user can now access WatchVideoEntryViewController by selecting a video entry in EntriesViewController. From there, they can watch the video entry or delete it. When a video entry is deleted, a notification is posted to EntriesViewModel that tells the view model to remove the deleted video entry from the videoEntries array. This results in realtime updates to the list of the user's video entries.