Add NoContentFoundView and Implement It in EntriesViewController
EntriesViewController now shows NoContentFoundView whenever text and/or video entries cannot be found. This has not been implemented for voice entries yet. To make this work, new view state options were added to EntriesViewState that are set when no video or text entries are found. This has also not been done for voice entries yet.